Consolidate pip reqs utils tests
